drft: skip test on linux ci

+++ Formula/d/drft.rb
@@ -13,20 +13,25 @@ class Drft < Formula
   end
 
   test do
-    (testpath/"diff.patch").write <<~EOS
-      --- a/file.txt
-      +++ b/file.txt
-      @@ -1 +1 @@
-      -Hello, world!
-      +Hello, Homebrew!
-    EOS
+    # failed with Linux CI, `code: 6, kind: Uncategorized, message: \"No such device or address\"`
+    return if OS.linux? && ENV["HOMEBREW_GITHUB_ACTIONS"]
 
-    output_log = testpath/"output.log"
-    pid = spawn bin/"drft", "diff.patch", [:out, :err] => output_log.to_s
-    sleep 1
-    assert_match "+Hello, Homebrew", output_log.read
-  ensure
-    Process.kill("TERM", pid)
-    Process.wait(pid)
+    begin
+      (testpath/"diff.patch").write <<~EOS
+        --- a/file.txt
+        +++ b/file.txt
+        @@ -1 +1 @@
+        -Hello, world!
+        +Hello, Homebrew!
+      EOS
+
+      output_log = testpath/"output.log"
+      pid = spawn bin/"drft", "diff.patch", [:out, :err] => output_log.to_s
+      sleep 1
+      assert_match "+Hello, Homebrew", output_log.read
+    ensure
+      Process.kill("TERM", pid)
+      Process.wait(pid)
+    end
   end
 end
